pyproject.toml
[project]
name = "petri-rewrite"
version = "0.1.0"
description = "Petri automated auditing framework — clean rewrite."
requires-python = ">=3.10"
dependencies = [
    "inspect-ai>=0.3.154",
    "inspect-swe",
    "agent-client-protocol>=0.8.1",
    "anthropic>=0.55.0",
    "pyyaml>=6.0",
    "anyio>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
